requirements-dev.txt
black==19.10b0
boto3==1.14.56
catboost==0.24.1
flake8==3.8.3
google-cloud-storage==1.31.0
isort==5.4.2
joblib==0.16.0
Keras==2.4.3
moto==1.3.14
onnxruntime==1.4.0
pylint==2.6.0
pytest==6.0.1
rope==0.17.0
scikit-learn==0.23.2
tensorflow==2.3.0
torch==1.6.0
torchvision==0.7.0
twine==3.2.0
xgboost==1.2.0